+++ app/src/main/java/com/peeranm/simpledictionary/feature_word_meaning/presentation/recent_searches/MainViewModel.kt
@@ -1,11 +1,8 @@
-package com.peeranm.simpledictionary.feature_word_meaning.presentation.main
+package com.peeranm.simpledictionary.feature_word_meaning.presentation.recent_searches
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.peeranm.simpledictionary.feature_word_meaning.data.remote.RetrofitInstance
-import com.peeranm.simpledictionary.feature_word_meaning.data.remote.dto.WordInfoDto
 import com.peeranm.simpledictionary.feature_word_meaning.model.WordInfo
-import com.peeranm.simpledictionary.feature_word_meaning.use_cases.GetWordInfoUseCase
 import com.peeranm.simpledictionary.feature_word_meaning.use_cases.WordInfoUseCases
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.*
