+++ app/src/main/res/layout/main_fragment.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".feature_word_meaning.presentation.main.MainFragment">
+    tools:context=".feature_word_meaning.presentation.recent_searches.MainFragment">
 
     <ImageView
         android:id="@+id/btnSearch"
