+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/mainFragment"
-        android:name="com.peeranm.simpledictionary.feature_word_meaning.presentation.main.MainFragment"
+        android:name="com.peeranm.simpledictionary.feature_word_meaning.presentation.recent_searches.MainFragment"
         android:label="main_fragment"
         tools:layout="@layout/main_fragment" >
         <action
